Wrap book page client logic in Suspense
